+++ packages/icons/scripts/helpers/capitalizeWord.js
@@ -1,0 +1,4 @@
+// eslint-disable-next-line no-magic-numbers
+const capitalizeWord = (word) => `${word[0].toUpperCase()}${word.slice(1)}`;
+
+module.exports = capitalizeWord;