+++ packages/icons/scripts/helpers/iconNameMapper.js
@@ -1,0 +1,5 @@
+const toCamelCase = require("./toCamelCase");
+
+const iconNameMapper = (iconName) => `case ${toCamelCase(iconName)} = "${iconName}"\n`;
+
+module.exports = iconNameMapper;