+++ packages/icons/scripts/helpers/toCamelCase.js
@@ -1,0 +1,8 @@
+const toCamelCaseMapper = require("./toCamelCaseMapper");
+
+const toCamelCase = (word) => word
+  .split("-")
+  .map(toCamelCaseMapper)
+  .join("");
+
+module.exports = toCamelCase;