+++ packages/icons/scripts/helpers/toCamelCaseMapper.js
@@ -1,0 +1,14 @@
+const capitalizeWord = require("./capitalizeWord");
+
+const toCamelCaseMapper = (item, index) => {
+
+  const FIRST_WORD = 0;
+
+  if (index === FIRST_WORD) {
+    return item;
+  }
+
+  return capitalizeWord(item);
+};
+
+module.exports = toCamelCaseMapper;