+++ packages/icons/scripts/helpers/toIosUnicode.js
@@ -1,0 +1,3 @@
+const toIosUnicode = (unicode) => `${unicode.replace("%u", "\\u{")}}`;
+
+module.exports = toIosUnicode;